Joined old modified version (LDAP group support) with fork of new snipe-it

+++ app/Http/Controllers/Assets/AssetsController.php
@@ -12,6 +12,7 @@ use App\Models\Company;
 use App\Models\Location;
 use App\Models\Setting;
 use App\Models\User;
+use App\Models\Group;
 use Auth;
 use Carbon\Carbon;
 use Intervention\Image\Facades\Image;
@@ -20,6 +21,7 @@ use Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Log;
 use Input;
 use League\Csv\Reader;
 use League\Csv\Statement;
@@ -79,11 +81,20 @@ class AssetsController extends Controller
      */
     public function create(Request $request)
     {
+        $user =  User::find(Auth::id());
+
+        if($user->isSuperUser()){
+            $userGroups = Group::pluck('name', 'id')->toArray();
+        }else{
+            $userGroups = $user->isAdminofGroup();
+        }
+        
         $this->authorize('create', Asset::class);
         $view = View::make('hardware/edit')
             ->with('statuslabel_list', Helper::statusLabelList())
             ->with('item', new Asset)
-            ->with('statuslabel_types', Helper::statusTypeList());
+            ->with('statuslabel_types', Helper::statusTypeList())
+            ->with('groups',$userGroups);
 
         if ($request->filled('model_id')) {
             $selected_model = AssetModel::find($request->input('model_id'));
@@ -207,6 +218,7 @@ class AssetsController extends Controller
         }
 
         if ($success) {
+            $asset->groups()->sync($request->input('groups'),false);
             // Redirect to the asset listing page
             return redirect()->route('hardware.index')
                 ->with('success', trans('admin/hardware/message.create.success'));
@@ -226,6 +238,17 @@ class AssetsController extends Controller
      */
     public function edit($assetId = null)
     {
+
+        $user =  User::find(Auth::id());
+
+        // $userGroups = $user->groups()->pluck('name', 'id');
+
+        if($user->isSuperUser()){
+            $userGroups = Group::pluck('name', 'id')->toArray();
+        }else{
+            $userGroups = $user->isAdminofGroup();
+        }
+
         if (!$item = Asset::find($assetId)) {
             // Redirect to the asset management page with error
             return redirect()->route('hardware.index')->with('error', trans('admin/hardware/message.does_not_exist'));
@@ -233,9 +256,20 @@ class AssetsController extends Controller
         //Handles company checks and permissions.
         $this->authorize($item);
 
-        return view('hardware/edit', compact('item'))
-            ->with('statuslabel_list', Helper::statusLabelList())
-            ->with('statuslabel_types', Helper::statusTypeList());
+        $assetGrp= $item->groups()->pluck('name', 'id')->toArray();
+            
+        $result = count(array_intersect($userGroups, $assetGrp));
+
+        if($result || $item->user_id == Auth::id()){
+
+            return view('hardware/edit', compact('item','assetGrp'))
+                    ->with('statuslabel_list', Helper::statusLabelList())
+                    ->with('statuslabel_types', Helper::statusTypeList())
+                    ->with('groups',$userGroups);
+
+        }else{
+            return redirect()->route('hardware.index')->with('error', 'You can not edit');
+        }
     }
 
 
@@ -368,6 +402,7 @@ class AssetsController extends Controller
 
 
         if ($asset->save()) {
+            $asset->groups()->sync($request->input('groups'),false);
             return redirect()->route("hardware.show", $assetId)
                 ->with('success', trans('admin/hardware/message.update.success'));
         }
@@ -393,21 +428,36 @@ class AssetsController extends Controller
 
         $this->authorize('delete', $asset);
 
-        DB::table('assets')
-            ->where('id', $asset->id)
-            ->update(array('assigned_to' => null));
+        $user =  User::find(Auth::id());
 
-        if ($asset->image) {
-            try  {
-                Storage::disk('public')->delete('assets'.'/'.$asset->image);
-            } catch (\Exception $e) {
-                \Log::debug($e);
-            }
+        if($user->isSuperUser()){
+            $userGroups = Group::pluck('name', 'id')->toArray();
+        }else{
+            $userGroups = $user->isAdminofGroup();
         }
 
-        $asset->delete();
+        $assetGrp= $asset->groups()->pluck('name', 'id')->toArray();
 
-        return redirect()->route('hardware.index')->with('success', trans('admin/hardware/message.delete.success'));
+        $result = count(array_intersect($userGroups, $assetGrp));
+        if($result || $asset->user_id == Auth::id()){
+            DB::table('assets')
+                ->where('id', $asset->id)
+                ->update(array('assigned_to' => null));
+
+            if ($asset->image) {
+                try  {
+                    Storage::disk('public')->delete('assets'.'/'.$asset->image);
+                } catch (\Exception $e) {
+                    \Log::debug($e);
+                }
+            }
+
+            $asset->delete();
+
+            return redirect()->route('hardware.index')->with('success', trans('admin/hardware/message.delete.success'));
+        }else{
+            return redirect()->route('hardware.index')->with('error', 'You can not delete');
+        }
     }
 
 
@@ -542,6 +592,14 @@ class AssetsController extends Controller
 
         $this->authorize('create', $asset_to_clone);
 
+        $user =  User::find(Auth::id());
+
+        if($user->isSuperUser()){
+            $userGroups = Group::pluck('name', 'id')->toArray();
+        }else{
+            $userGroups = $user->isAdminofGroup();
+        }
+
         $asset = clone $asset_to_clone;
         $asset->id = null;
         $asset->asset_tag = '';
@@ -551,7 +609,8 @@ class AssetsController extends Controller
         return view('hardware/edit')
             ->with('statuslabel_list', Helper::statusLabelList())
             ->with('statuslabel_types', Helper::statusTypeList())
-            ->with('item', $asset);
+            ->with('item', $asset)
+            ->with('groups',$userGroups);
     }
 
     /**
